moved code into homescreen and separated the available_event_widget.dart
